app/requirements.txt
langchain==0.0.138
requests==2.28.2
sentence-transformers==2.2.2
streamlit==1.24.1
feedparser==6.0.10
qdrant-client==1.7.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
InstructorEmbedding==1.0.1